+++ app/Type/PlainObject.ts
@@ -1,6 +1,8 @@
+export type Primitives = string | number | boolean;
+
 /**
  * One-level depth objects with simple values.
  */
 export type PlainObject = {
-  [key: string]: string | number | boolean;
+  [key: string]: Primitives | Array<Primitives>;
 }